Moved connector.py to the components module
